fix(db): type NodeEvent.exit_code as i64 to match its source column

+++ crates/db/src/models/event.rs
@@ -45,7 +45,7 @@ pub enum NodeEvent {
         attempt_id: Uuid,
         execution_process_id: Uuid,
         executor: String,
-        exit_code: i32,
+        exit_code: i64,
     },
     /// An attempt failed.
     AttemptFailed {
